+++ app/src/main/java/com/pruden/habits/adapters/RegistroNumericoAdapter.kt
@@ -8,7 +8,7 @@ import com.pruden.habits.R
 import com.pruden.habits.adapters.listeners.OnClickNumericoRegistro
 import com.pruden.habits.databinding.ItemRegistroNumericoBinding
 
-class RegistroNumericoAdapter (val listaRegistros: MutableList<Int>, val unidad: String, val listener: OnClickNumericoRegistro)
+class RegistroNumericoAdapter (val listaRegistros: MutableList<Float>, val unidad: String, val listener: OnClickNumericoRegistro)
     : RecyclerView.Adapter<RegistroNumericoAdapter.ViewHolder>() {
     inner class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         val binding = ItemRegistroNumericoBinding.bind(view)
