+++ app/src/main/java/com/pruden/habits/adapters/listeners/OnClickNumericoRegistro.kt
@@ -5,6 +5,6 @@ import android.widget.TextView
 import com.google.android.material.textfield.TextInputEditText
 
 interface OnClickNumericoRegistro {
-    fun onClickNumericoRegistro(textField : TextView ,valor : Int, unidad: String)
+    fun onClickNumericoRegistro(textField : TextView ,valor : Float, unidad: String)
 
 }
